Move endpoint tests
